mpc-connect/requirements.txt
azure-functions==1.14.0
ipython==8.13.1
planetary-computer==0.5.1
pystac==1.7.3
pystac-client==0.6.1
geopandas==0.13.0
Fiona==1.9.4.post1
pyproj==3.5.0
Rtree==1.0.1
shapely==2.0.1
matplotlib-inline==0.1.6
matplotlib
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pyspark
Pillow==9.5.0